Hidden Inputs should never be localized.
When making localized=True the default we ran into the problem that Hidden Inputs
are also localized, even though they may contain pk values and localization is not
very interesting for something that is, by definition, hidden.

Other fields that may work with PKs (select boxes etc.) already turn off localization
of the pk value.

So the chosen solution here was simply to patch the HiddenInput as the only known
non-working case.