Don't localize boolean values for hidden_inputs
(i.e. as T.rue and Fa.lse because of the thousands_separator)

6c53a81 only fixed localize(), but not localize_input()
It appears to me that this was simply an oversight at the time.

It seems that a new call-order for fields somewhere between Django 1.4
and Django 1.6 triggers triggers this bug again, but this time on the inputs
and only when rendering fields as_hidden.

This is the "graft"/"cherry pick" of f0e39c83ec91 which does the same
thing on Django 1.6